fix #All: Completed all duties
