+++ build.gradle
@@ -1,37 +1,54 @@
 /*
  * This file was generated by the Gradle 'init' task.
+ *
+ * This generated file contains a sample Java Library project to get you started.
+ * For more details take a look at the Java Libraries chapter in the Gradle
+ * User Manual available at https://docs.gradle.org/6.0/userguide/java_library_plugin.html
  */
 
 plugins {
-	id "com.github.roroche.plantuml" version "1.0.2"
+    // Apply the java-library plugin to add support for Java Library
+    id 'java-library'
     id 'java'
-    id 'maven-publish'
+    id 'application'
+    id "com.github.roroche.plantuml" version "1.0.2"
 }
 
 repositories {
-    mavenLocal()
-    maven {
-        url = uri('http://repo.maven.apache.org/maven2')
-    }
+    // Use jcenter for resolving dependencies.
+    // You can declare any Maven/Ivy/file repository here.
+    jcenter()
 }
 
 dependencies {
-    runtimeOnly 'org.junit.jupiter:junit-jupiter-api:5.6.0'
+    // This dependency is exported to consumers, that is to say found on their compile classpath.
+    api 'org.apache.commons:commons-math3:3.6.1'
+
+    // This dependency is used internally, and not exposed to consumers on their own compile classpath.
+	implementation 'com.google.guava:guava:28.2-jre'
+	
+    // Use JUnit test framework
     testImplementation 'junit:junit:4.12'
+    
+    
 }
 
-group = 'org.blagnac.cpoa'
-version = '1.0.0-SNAPSHOT'
-sourceCompatibility = '1.8'
-
-publishing {
-    publications {
-        maven(MavenPublication) {
-            from(components.java)
-        }
-    }
+application {
+    // Define the main class for the application.
+    mainClassName = 'tp3.PizzaTestDrive'
 }
 
-tasks.withType(JavaCompile) {
-    options.encoding = 'UTF-8'
+jar {
+    destinationDirectory.set(file("target/"))
+    baseName 'pizzeria'
+
+    manifest { 
+        attributes 'Main-Class': 'tp3.PizzaTestDrive' 
+    } 
+
+}
+
+classDiagram { // (1)
+    packageName = "tp3" // (2)
+    outputFile = project.file('docs/TP3.plantuml') // (3)
 }
