+++ docs/TP3.plantuml
@@ -1,60 +1,61 @@
 @startuml
 
-abstract class Pizza {
+abstract class "Pizza" {
   name : String
   crust : String
   sauce : String
-  garnitures : ArrayList
+  garnishes : ArrayList<String>
   bake()
   cut()
   prepare()
   wrap()
 }
 
-class PizzaCheeseStyleBrest
+class "PizzaCheeseStyleBrest"
 
-class PizzaCheeseStyleStrasbourg {
+class "PizzaCheeseStyleStrasbourg" {
   cut()
 }
 
-abstract class PizzaFactory {
-  create(String) : Pizza
-}
+abstract class "PizzaFactory"
 
-class PizzaFactoryBrest {
-  create(String) : Pizza
-}
+class "PizzaFactoryBrest"
 
-class PizzaFactoryStrasbourg {
-  create(String) : Pizza
-}
+class "PizzaFactoryStrasbourg"
 
-class PizzaTestDrive
+class "PizzaGrecqueStyleBrest"
 
-abstract class Pizzeria {
-  pizzaFactory : PizzaFactory
+class "PizzaGrecqueStyleStrasbourg"
+
+class "PizzaPoivronStyleBrest"
+
+class "PizzaPoivronStyleStrasbourg"
+
+class "PizzaTestDrive"
+
+abstract class "Pizzeria" {
   orderPizza(String) : Pizza
 }
 
-class PizzeriaBrest
+class "PizzeriaBrest"
 
-class PizzeriaFactory {
+class "PizzeriaFactory" {
   create(String) : Pizzeria
 }
 
-class PizzeriaStrasbourg
+class "PizzeriaStrasbourg"
 
-PizzaFactoryBrest --> Pizza : use
-PizzaFactory --> Pizza : use
-PizzaFactory <|-- PizzaFactoryBrest
-PizzaFactory <|-- PizzaFactoryStrasbourg
-PizzaFactoryStrasbourg --> Pizza : use
-Pizza <|-- PizzaCheeseStyleBrest
-Pizza <|-- PizzaCheeseStyleStrasbourg
-PizzeriaFactory --> Pizzeria : use
-Pizzeria --> Pizza : use
-Pizzeria --> "1" PizzaFactory : pizzaFactory
-Pizzeria <|-- PizzeriaBrest
-Pizzeria <|-- PizzeriaStrasbourg
+"PizzaFactory" <|-- "PizzaFactoryBrest"
+"PizzaFactory" <|-- "PizzaFactoryStrasbourg"
+"Pizza" <|-- "PizzaCheeseStyleBrest"
+"Pizza" <|-- "PizzaCheeseStyleStrasbourg"
+"Pizza" <|-- "PizzaGrecqueStyleBrest"
+"Pizza" <|-- "PizzaGrecqueStyleStrasbourg"
+"Pizza" <|-- "PizzaPoivronStyleBrest"
+"Pizza" <|-- "PizzaPoivronStyleStrasbourg"
+"PizzeriaFactory" --> "Pizzeria" : use
+"Pizzeria" --> "Pizza" : use
+"Pizzeria" <|-- "PizzeriaBrest"
+"Pizzeria" <|-- "PizzeriaStrasbourg"
 
 @enduml
