+++ src/main/java/tp3/PizzaCheeseStyleBrest.java
@@ -1,9 +1,11 @@
+package tp3;
 import java.util.ArrayList;
 
-public class PizzaPoivronStyleStrasbourg extends Pizza{
-	public PizzaPoivronStyleStrasbourg() {
+public class PizzaCheeseStyleBrest extends Pizza {
+	
+	public PizzaCheeseStyleBrest() {
 		this.garnishes = new ArrayList<String>();
-        this.garnishes.add("Poivrons");
+        this.garnishes.add("Parmigiano reggiano");
         System.out.println("Preparation of Pizza with Brest style sauce and cheese");
 	}
