+++ src/main/java/tp3/PizzaPoivronStyleStrasbourg.java
@@ -1,10 +1,10 @@
+package tp3;
 import java.util.ArrayList;
 
-public class PizzaCheeseStyleBrest extends Pizza {
-	
-	public PizzaCheeseStyleBrest() {
+public class PizzaPoivronStyleStrasbourg extends Pizza{
+	public PizzaPoivronStyleStrasbourg() {
 		this.garnishes = new ArrayList<String>();
-        this.garnishes.add("Parmigiano reggiano");
+        this.garnishes.add("Poivrons");
         System.out.println("Preparation of Pizza with Brest style sauce and cheese");
 	}
