+++ src/main/java/tp3/Pizzeria.java
@@ -1,3 +1,4 @@
+package tp3;
 public abstract class Pizzeria {
 	protected Pizza createPizza (String type) {
